move methods from wal_generator.go to wal_test.go
they are only used there anyways
having exported method only for tests is meh
